NCKH/IT2FCM(Not mine)/CoreFCM/requirements.txt
affine==2.4.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
filelock==3.16.1
fsspec==2024.9.0
imageio==2.36.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
MarkupSafe==3.0.1
mpmath==1.3.0
networkx==3.4
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
rasterio==1.4.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.1.0
shuffle==0.9
six==1.16.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
torch==2.4.1
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.2
ucimlrepo==0.0.7
Unidecode==1.3.8
wheel==0.44.0
XlsxWriter==3.2.0
